requirements.txt
adal==1.0.2
azure-batch==4.1.3
azure-cosmosdb-table==1.0.4
azure-keyvault==1.0.0
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-compute==4.0.1
azure-mgmt-network==2.0.0
azure-mgmt-resource==2.0.0
azure-mgmt-storage==2.0.0rc4
azure-storage-blob==1.3.1
azure-storage-file==1.3.1
azure-storage-queue==1.3.0
blobxfer==1.3.1
click==6.7
future==0.16.0
futures==3.2.0; python_version < '3'
keyrings.alt==3.1
msrest==0.5.4
msrestazure==0.4.34
pathlib2==2.3.2; python_version < '3.5'
python-dateutil==2.7.3
requests==2.19.1
ruamel.yaml==0.15.46
scandir==1.7; python_version < '3.5'
